requirements.txt
fastapi
uvicorn[standard]
pillow
transformers
torch
SpeechRecognition
pyttsx3
numpy
scipy
